bug fix on css and modulize js file
