+++ src/style.css
@@ -8,9 +8,7 @@ html {
   -webkit-text-size-adjust: 100%;
 }
 
-*,
-*::before,
-*::after {
+* {
   box-sizing: inherit;
   margin: 0;
   padding: 0;
@@ -61,15 +59,13 @@ p {
   box-shadow: 0 10px 30px rgba(65, 72, 86, 0.05);
   max-width: 25rem;
   width: 100%;
-  padding: 1rem;
   margin: 0 auto;
 }
 
 .clear {
   border: none;
   max-width: 31rem;
-  width: 109%;
-  margin: -18px -16px;
+  width: 100%;
   height: 2.3rem;
 }
